Separate getTrafficFlows in other controller
