+++ app/api/controllers.go
@@ -26,19 +26,6 @@ type Api struct {
 	*gin.Engine
 }
 
-func (api *Api) GetTraffic(c *gin.Context) {
-	traffic, err := api.TrafficSearcher.GetTrafficFlows()
-	if err != nil {
-		fmt.Println(err)
-		c.JSON(500, gin.H{"data": "error"})
-		c.AbortWithStatus(http.StatusInternalServerError)
-		return
-	}
-	c.Header("Access-Control-Allow-Origin", "*") //There is a vuln here, that's only for testing purpose.
-	c.Header("Access-Control-Allow-Methods", "GET")
-	c.JSON(http.StatusOK, gin.H{"data": traffic})
-}
-
 func (api *Api) GetActiveFlowsPerDestination(c *gin.Context) {
 	activeFlows, err := api.TrafficBytesParser.GetBytesPerDestination()
 	if err != nil {
