+++ app/api/get_traffic_flows_handler.go
@@ -1,0 +1,21 @@
+package api
+
+import (
+	"fmt"
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+)
+
+func (api *Api) GetTraffic(c *gin.Context) {
+	traffic, err := api.TrafficSearcher.GetTrafficFlows()
+	if err != nil {
+		fmt.Println(err)
+		c.JSON(500, gin.H{"data": "error"})
+		c.AbortWithStatus(http.StatusInternalServerError)
+		return
+	}
+	c.Header("Access-Control-Allow-Origin", "*") //There is a vuln here, that's only for testing purpose.
+	c.Header("Access-Control-Allow-Methods", "GET")
+	c.JSON(http.StatusOK, gin.H{"data": traffic})
+}
